requirements.txt
python-dotenv==1.2.1
openai==2.17.0
requests==2.32.5
streamlit==1.55.0